Backporting alias bug fixes

+++ web/concrete/src/Multilingual/Page/Section/Section.php
@@ -360,6 +360,9 @@ class Section extends Page
 
     public static function registerPage($page)
     {
+        if ($page->isAlias()) {
+            return;
+        }
         if (\Core::make('multilingual/detector')->isEnabled()) {
             $db = Database::get();
             $ms = static::getBySectionOfSite($page);
@@ -395,6 +398,9 @@ class Section extends Page
 
     public static function unregisterPage($page)
     {
+        if ($page->isAlias()) {
+            return;
+        }
         $db = Database::get();
         $db->Execute('delete from MultilingualSections where cID = ?', array($page->getCollectionID()));
         $db->Execute('delete from MultilingualPageRelations where cID = ?', array($page->getCollectionID()));
@@ -402,6 +408,10 @@ class Section extends Page
 
     public static function registerMove($page, $oldParent, $newParent)
     {
+        if ($page->isAlias()) {
+            return;
+        }
+
         if (static::isMultilingualSection($newParent)) {
             $ms = static::getByID($newParent->getCollectionID());
         } else {
